+++ Example-Swift/JudoKitSwiftExample/ViewController.swift
@@ -209,7 +209,7 @@ class ViewController: UIViewController, PKPaymentAuthorizationViewControllerDele
             viewController.response = response
             self.navigationController?.pushViewController(viewController, animated: true)
             }, errorHandler: { (error) -> () in
-                if error == JudoError.UserDidCancel {
+                if error.judoCode == .UserDidCancel {
                     self.dismissViewControllerAnimated(true, completion: nil)
                 }
                 // handle other errors that may encounter
@@ -234,7 +234,7 @@ class ViewController: UIViewController, PKPaymentAuthorizationViewControllerDele
             viewController.response = response
             self.navigationController?.pushViewController(viewController, animated: true)
             }, errorHandler: { (error) -> () in
-                if error == JudoError.UserDidCancel {
+                if error.judoCode == .UserDidCancel {
                     self.dismissViewControllerAnimated(true, completion: nil)
                 }
                 // handle other errors that may encounter
@@ -254,7 +254,7 @@ class ViewController: UIViewController, PKPaymentAuthorizationViewControllerDele
                 self.paymentToken = transactionData.paymentToken()
             }
             }, errorHandler: { (error) -> () in
-                if error == JudoError.UserDidCancel {
+                if error.judoCode == .UserDidCancel {
                     self.dismissViewControllerAnimated(true, completion: nil)
                 }
                 // handle other errors that may encounter
@@ -279,7 +279,7 @@ class ViewController: UIViewController, PKPaymentAuthorizationViewControllerDele
                 viewController.response = response
                 self.navigationController?.pushViewController(viewController, animated: true)
                 }, errorHandler: { (error) -> () in
-                    if error == JudoError.UserDidCancel {
+                    if error.judoCode == .UserDidCancel {
                         self.dismissViewControllerAnimated(true, completion: nil)
                     }
                     // handle other errors that may encounter
@@ -309,7 +309,7 @@ class ViewController: UIViewController, PKPaymentAuthorizationViewControllerDele
                 viewController.response = response
                 self.navigationController?.pushViewController(viewController, animated: true)
                 }, errorHandler: { (error) -> () in
-                    if error == JudoError.UserDidCancel {
+                    if error.judoCode == .UserDidCancel {
                         self.dismissViewControllerAnimated(true, completion: nil)
                     }
                     // handle other errors that may encounter