+++ Source/CardInputField.swift
@@ -75,7 +75,7 @@ public class CardInputField: JudoPayInputField {
                 self.delegate?.cardInput(self, didFindValidNumber: textField.text!)
                 self.dismissError()
             } else {
-                self.delegate?.cardInput(self, error: JudoError.InvalidCardNumber)
+                self.delegate?.cardInput(self, error: JudoError(.InvalidCardNumber))
             }
         }