+++ Source/String+Card.swift
@@ -68,9 +68,9 @@ public extension String {
         if strippedSelf.characters.count == 0 {
             return ""
         } else if strippedSelf.characters.count > Card.maximumLength {
-            throw JudoError.CardLengthMismatchError
+            throw JudoError(.CardLengthMismatchError)
         } else if !strippedSelf.isNumeric() {
-            throw JudoError.InvalidEntry
+            throw JudoError(.InvalidEntry)
         }
         
         // make sure to only check validity for the necessary networks
@@ -89,7 +89,7 @@ public extension String {
         
         if patterns.count == 0 {
             // if no patterns are left - the entered number is invalid
-            throw JudoError.InvalidCardNumber
+            throw JudoError(.InvalidCardNumber)
         }
         
         // retrieve the shortest pattern that is left and start moving the characters across